# Don't print symbols files as compiler diagnostic notes